[glyf] Split composite-iter
